+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/Dynatrace/openkit-java/compare/v1.4.2...release/1.4)
 
+### Changed
+- Fix sending of session number in web request tracer tag to honor the data collection level.
+  The session number will only be sent with data collection level 'User Behavior'.
+
 ## 1.4.2 [Release date: 2019-07-17]
 [GitHub Releases](https://github.com/Dynatrace/openkit-java/releases/tag/v1.4.2)